MOBILE-3833 core: Add cookie to force logs
